requirements.txt
pygithub==1.55
tqdm==4.60.0
pandas==1.2.4
gitpython==3.1.14
lxml==4.6.3
